+++ xskillscore/tests/test_skipna_functionality.py
@@ -54,6 +54,33 @@ def weights():
     return da
 
 
+@pytest.fixture
+def a_3d():
+    da = xr.DataArray(
+        np.repeat(np.array([[0, 1, 2], [4, 5, 6], [7, 8, 9]]), 3).reshape(3, 3, 3),
+        coords=[
+            pd.date_range("1/1/2000", "1/3/2000", freq="D"),
+            np.linspace(-89.5, 89.5, 3),
+            np.linspace(-179.5, 179.5, 3),
+        ],
+        dims=["time", "lat", "lon"],
+    )
+    return da
+
+
+@pytest.fixture
+def b_3d(a_3d):
+    return a_3d.copy()
+
+
+@pytest.fixture
+def weights_3d(a_3d):
+    weights = np.cos(np.deg2rad(a_3d.lat))
+    _, weights = xr.broadcast(a_3d, weights)
+    weights = weights.isel(time=0)
+    return weights
+
+
 def drop_nans(a, b, weights=None, dim="time"):
     """
     Masks a and b where they have pairwise nans.
@@ -94,3 +121,15 @@ def test_skipna_returns_nan_when_false(a, b, metric):
     and skipna is False."""
     res = metric(a, b, "time", skipna=False)
     assert np.isnan(res).all()
+
+
+@pytest.mark.parametrize("metric", WEIGHTED_METRICS)
+def test_skipna_broadcast_weights_assignment_destination(
+    a_3d, b_3d, weights_3d, metric
+):
+    """Tests that 'assignment destination is read-only' is not raised
+    https://github.com/raybellwaves/xskillscore/issues/79"""
+    a_3d_nan = a_3d.where(a_3d > 0)
+    b_3d_nan = b_3d.where(b_3d > 0)
+    res = metric(a_3d_nan, b_3d_nan, ["lat", "lon"], weights=weights_3d, skipna=True)
+    assert all([not np.isnan(r) for r in res])